moved all pouchdb tests to wrong suite
